bootstrap tags updated, data structure change required for the rest

+++ public/src/js/plugins/bootstrap-tags.js
@@ -56,6 +56,10 @@
         },
         onLoadSuggestions: function(values) {
             return values;
+        },
+        onDuplicate: null,
+        onBeforeRemove: function(pill) {
+            return true;
         }
     }
 
@@ -231,13 +235,19 @@
         });
 
         if(unique) {
-            var color = $(pills_list.children()[0]).css('background-color');
-            unique.stop().animate({"backgroundColor": $self.options.double_hilight}, 100, 'swing', function() {
-                unique.stop().animate({"backgroundColor": color}, 100, 'swing', function(){
-                   unique.css('background-color', '');
+            if(!$self.options.onDuplicate){
+                var color = $(pills_list.children()[0]).css('background-color');
+                unique.stop().animate({"backgroundColor": $self.options.double_hilight}, 100, 'swing', function() {
+                    unique.stop().animate({"backgroundColor": color}, 100, 'swing', function(){
+                        unique.css('background-color', '');
+                    });
                 });
-            });
-            return false;
+                return false;
+            } else {
+                if($self.options.onDuplicate(unique, value) != true) {
+                    return false;
+                }
+            }
         }
 
         if(value.url) {
@@ -290,13 +300,21 @@
 
     Tags.prototype.removeTag = function(tag) {
         var $self = this;
-        $(tag).closest('[data-tag-id]').animate({width: 0, "padding-right": 0, "padding-left": 0}, 200, 'swing', function() {
+        var $tag = $(tag).closest('[data-tag-id]');
+
+        if($self.options.remove_url) {
+            $.ajax({
+                dataType: 'json', type: 'post', async: false, url: $self.options.remove_url, data: {id: $tag.data('tag-id')}
+            });
+        }
+
+        if($self.options.onBeforeRemove($tag) === false) {
+            return;
+        }
+
+        $tag.animate({width: 0, "padding-right": 0, "padding-left": 0}, 200, 'swing', function() {
             var $this = $(this);
-            if($self.options.remove_url) {
-                $.ajax({
-                    dataType: 'json', type: 'post', async: false, url: $self.options.remove_url, data: {id: $this.data('tag-id')}
-                });
-            }
+ 
             $self.options.onRemove($this);
             $this.remove();
         });
